feat: extract hashing utils
